adds detail view animations
